requirements.txt
python-fasthtml>=0.12.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
